+++ src/cs3500/animator/model/commands/AddRectCmd.java
@@ -27,7 +27,7 @@ public class AddRectCmd extends InstantCmd {
    * @param color  the color of the new Rectangle.
    * @param tick   the tick when this command triggers.
    */
-  public AddRectCmd(String name, float x, float y, int width, int height, Color color, int tick) {
+  public AddRectCmd(String name, double x, double y, int width, int height, Color color, int tick) {
     super(name, tick);
     this.rect = new Rectangle(name, x, y, width, height, color);
   }
