+++ src/cs3500/animator/model/interfaces/Movable.java
@@ -14,7 +14,7 @@ public interface Movable {
   /**
    * Places this object at the given coordinates.
    */
-  void place(float x, float y);
+  void place(double x, double y);
 
   /**
    * Places this object at the given coordinates.
@@ -29,7 +29,7 @@ public interface Movable {
    * @param dx the change in x.
    * @param dy the change in y.
    */
-  void move(float dx, float dy);
+  void move(double dx, double dy);
 
   /**
    * Get the current position of this object.
